v 1.0.0.0 Se volvio a hacer la capa de persistencia, se agrego funciones de creacion de compra

+++ src/main/java/cunori/kardex/controller/CompraJpaController.java
@@ -4,17 +4,22 @@
  */
 package cunori.kardex.controller;
 
+import cunori.kardex.controller.exceptions.IllegalOrphanException;
 import cunori.kardex.controller.exceptions.NonexistentEntityException;
 import cunori.kardex.controller.exceptions.PreexistingEntityException;
 import cunori.kardex.dao.Compra;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import cunori.kardex.dao.KardexPEPS;
+import java.util.ArrayList;
+import java.util.Collection;
+import cunori.kardex.dao.DetalleCompra;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -32,11 +37,47 @@ public class CompraJpaController implements Serializable {
     }
 
     public void create(Compra compra) throws PreexistingEntityException, Exception {
+        if (compra.getKardexPEPSCollection() == null) {
+            compra.setKardexPEPSCollection(new ArrayList<KardexPEPS>());
+        }
+        if (compra.getDetalleCompraCollection() == null) {
+            compra.setDetalleCompraCollection(new ArrayList<DetalleCompra>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Collection<KardexPEPS> attachedKardexPEPSCollection = new ArrayList<KardexPEPS>();
+            for (KardexPEPS kardexPEPSCollectionKardexPEPSToAttach : compra.getKardexPEPSCollection()) {
+                kardexPEPSCollectionKardexPEPSToAttach = em.getReference(kardexPEPSCollectionKardexPEPSToAttach.getClass(), kardexPEPSCollectionKardexPEPSToAttach.getId());
+                attachedKardexPEPSCollection.add(kardexPEPSCollectionKardexPEPSToAttach);
+            }
+            compra.setKardexPEPSCollection(attachedKardexPEPSCollection);
+            Collection<DetalleCompra> attachedDetalleCompraCollection = new ArrayList<DetalleCompra>();
+            for (DetalleCompra detalleCompraCollectionDetalleCompraToAttach : compra.getDetalleCompraCollection()) {
+                detalleCompraCollectionDetalleCompraToAttach = em.getReference(detalleCompraCollectionDetalleCompraToAttach.getClass(), detalleCompraCollectionDetalleCompraToAttach.getId());
+                attachedDetalleCompraCollection.add(detalleCompraCollectionDetalleCompraToAttach);
+            }
+            compra.setDetalleCompraCollection(attachedDetalleCompraCollection);
             em.persist(compra);
+            for (KardexPEPS kardexPEPSCollectionKardexPEPS : compra.getKardexPEPSCollection()) {
+                Compra oldIdCompraOfKardexPEPSCollectionKardexPEPS = kardexPEPSCollectionKardexPEPS.getIdCompra();
+                kardexPEPSCollectionKardexPEPS.setIdCompra(compra);
+                kardexPEPSCollectionKardexPEPS = em.merge(kardexPEPSCollectionKardexPEPS);
+                if (oldIdCompraOfKardexPEPSCollectionKardexPEPS != null) {
+                    oldIdCompraOfKardexPEPSCollectionKardexPEPS.getKardexPEPSCollection().remove(kardexPEPSCollectionKardexPEPS);
+                    oldIdCompraOfKardexPEPSCollectionKardexPEPS = em.merge(oldIdCompraOfKardexPEPSCollectionKardexPEPS);
+                }
+            }
+            for (DetalleCompra detalleCompraCollectionDetalleCompra : compra.getDetalleCompraCollection()) {
+                Compra oldIdCompraOfDetalleCompraCollectionDetalleCompra = detalleCompraCollectionDetalleCompra.getIdCompra();
+                detalleCompraCollectionDetalleCompra.setIdCompra(compra);
+                detalleCompraCollectionDetalleCompra = em.merge(detalleCompraCollectionDetalleCompra);
+                if (oldIdCompraOfDetalleCompraCollectionDetalleCompra != null) {
+                    oldIdCompraOfDetalleCompraCollectionDetalleCompra.getDetalleCompraCollection().remove(detalleCompraCollectionDetalleCompra);
+                    oldIdCompraOfDetalleCompraCollectionDetalleCompra = em.merge(oldIdCompraOfDetalleCompraCollectionDetalleCompra);
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findCompra(compra.getId()) != null) {
@@ -50,12 +91,73 @@ public class CompraJpaController implements Serializable {
         }
     }
 
-    public void edit(Compra compra) throws NonexistentEntityException, Exception {
+    public void edit(Compra compra) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Compra persistentCompra = em.find(Compra.class, compra.getId());
+            Collection<KardexPEPS> kardexPEPSCollectionOld = persistentCompra.getKardexPEPSCollection();
+            Collection<KardexPEPS> kardexPEPSCollectionNew = compra.getKardexPEPSCollection();
+            Collection<DetalleCompra> detalleCompraCollectionOld = persistentCompra.getDetalleCompraCollection();
+            Collection<DetalleCompra> detalleCompraCollectionNew = compra.getDetalleCompraCollection();
+            List<String> illegalOrphanMessages = null;
+            for (KardexPEPS kardexPEPSCollectionOldKardexPEPS : kardexPEPSCollectionOld) {
+                if (!kardexPEPSCollectionNew.contains(kardexPEPSCollectionOldKardexPEPS)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain KardexPEPS " + kardexPEPSCollectionOldKardexPEPS + " since its idCompra field is not nullable.");
+                }
+            }
+            for (DetalleCompra detalleCompraCollectionOldDetalleCompra : detalleCompraCollectionOld) {
+                if (!detalleCompraCollectionNew.contains(detalleCompraCollectionOldDetalleCompra)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain DetalleCompra " + detalleCompraCollectionOldDetalleCompra + " since its idCompra field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<KardexPEPS> attachedKardexPEPSCollectionNew = new ArrayList<KardexPEPS>();
+            for (KardexPEPS kardexPEPSCollectionNewKardexPEPSToAttach : kardexPEPSCollectionNew) {
+                kardexPEPSCollectionNewKardexPEPSToAttach = em.getReference(kardexPEPSCollectionNewKardexPEPSToAttach.getClass(), kardexPEPSCollectionNewKardexPEPSToAttach.getId());
+                attachedKardexPEPSCollectionNew.add(kardexPEPSCollectionNewKardexPEPSToAttach);
+            }
+            kardexPEPSCollectionNew = attachedKardexPEPSCollectionNew;
+            compra.setKardexPEPSCollection(kardexPEPSCollectionNew);
+            Collection<DetalleCompra> attachedDetalleCompraCollectionNew = new ArrayList<DetalleCompra>();
+            for (DetalleCompra detalleCompraCollectionNewDetalleCompraToAttach : detalleCompraCollectionNew) {
+                detalleCompraCollectionNewDetalleCompraToAttach = em.getReference(detalleCompraCollectionNewDetalleCompraToAttach.getClass(), detalleCompraCollectionNewDetalleCompraToAttach.getId());
+                attachedDetalleCompraCollectionNew.add(detalleCompraCollectionNewDetalleCompraToAttach);
+            }
+            detalleCompraCollectionNew = attachedDetalleCompraCollectionNew;
+            compra.setDetalleCompraCollection(detalleCompraCollectionNew);
             compra = em.merge(compra);
+            for (KardexPEPS kardexPEPSCollectionNewKardexPEPS : kardexPEPSCollectionNew) {
+                if (!kardexPEPSCollectionOld.contains(kardexPEPSCollectionNewKardexPEPS)) {
+                    Compra oldIdCompraOfKardexPEPSCollectionNewKardexPEPS = kardexPEPSCollectionNewKardexPEPS.getIdCompra();
+                    kardexPEPSCollectionNewKardexPEPS.setIdCompra(compra);
+                    kardexPEPSCollectionNewKardexPEPS = em.merge(kardexPEPSCollectionNewKardexPEPS);
+                    if (oldIdCompraOfKardexPEPSCollectionNewKardexPEPS != null && !oldIdCompraOfKardexPEPSCollectionNewKardexPEPS.equals(compra)) {
+                        oldIdCompraOfKardexPEPSCollectionNewKardexPEPS.getKardexPEPSCollection().remove(kardexPEPSCollectionNewKardexPEPS);
+                        oldIdCompraOfKardexPEPSCollectionNewKardexPEPS = em.merge(oldIdCompraOfKardexPEPSCollectionNewKardexPEPS);
+                    }
+                }
+            }
+            for (DetalleCompra detalleCompraCollectionNewDetalleCompra : detalleCompraCollectionNew) {
+                if (!detalleCompraCollectionOld.contains(detalleCompraCollectionNewDetalleCompra)) {
+                    Compra oldIdCompraOfDetalleCompraCollectionNewDetalleCompra = detalleCompraCollectionNewDetalleCompra.getIdCompra();
+                    detalleCompraCollectionNewDetalleCompra.setIdCompra(compra);
+                    detalleCompraCollectionNewDetalleCompra = em.merge(detalleCompraCollectionNewDetalleCompra);
+                    if (oldIdCompraOfDetalleCompraCollectionNewDetalleCompra != null && !oldIdCompraOfDetalleCompraCollectionNewDetalleCompra.equals(compra)) {
+                        oldIdCompraOfDetalleCompraCollectionNewDetalleCompra.getDetalleCompraCollection().remove(detalleCompraCollectionNewDetalleCompra);
+                        oldIdCompraOfDetalleCompraCollectionNewDetalleCompra = em.merge(oldIdCompraOfDetalleCompraCollectionNewDetalleCompra);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -73,7 +175,7 @@ public class CompraJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -84,6 +186,24 @@ public class CompraJpaController implements Serializable {
                 compra.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The compra with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<KardexPEPS> kardexPEPSCollectionOrphanCheck = compra.getKardexPEPSCollection();
+            for (KardexPEPS kardexPEPSCollectionOrphanCheckKardexPEPS : kardexPEPSCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Compra (" + compra + ") cannot be destroyed since the KardexPEPS " + kardexPEPSCollectionOrphanCheckKardexPEPS + " in its kardexPEPSCollection field has a non-nullable idCompra field.");
+            }
+            Collection<DetalleCompra> detalleCompraCollectionOrphanCheck = compra.getDetalleCompraCollection();
+            for (DetalleCompra detalleCompraCollectionOrphanCheckDetalleCompra : detalleCompraCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Compra (" + compra + ") cannot be destroyed since the DetalleCompra " + detalleCompraCollectionOrphanCheckDetalleCompra + " in its detalleCompraCollection field has a non-nullable idCompra field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(compra);
             em.getTransaction().commit();
